pyproject.toml
[tool.poetry]
name = "maestro-api"
version = "0.1.0"
description = ""
authors = ["William R. Stetar <kebekad673@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
